Fix session incompatibility with XH 1.7
As of CMSimple_XH 1.7.0 the session uses a non-standard name, which
breaks compatibility with elFinder_xh 1.0.5. We fix this by checking
whether cmsimple/.sessionname exist, and if so we use its content to
set the appropriate session name before actually starting the session.

We also cater to `$_SESSION['xh_password']` now holding a string
instead of an array as before.